requirements.txt
flask
pillow
tensorflow
imutils
webcolors